minor code changes

+++ src/includes/class-activator.php
@@ -31,21 +31,27 @@ class Bonaire_Activator {
 	/**
 	 * Instantiates the class responsible for activating the plugin.
 	 *
-	 * @since0.9.6
+	 * @since 0.9.6
 	 * @return void
 	 */
 	public function activate() {
 		
-		// Adds default settings data if there are no options stored yet.
-		if ( false === get_option( 'bonaire_options' ) || '1' === get_transient( 'bonaire_reset_settings' ) ) {
-			delete_option( 'bonaire_options' );
-			$Bonaire_Options = new AdminIncludes\Bonaire_Options( 'bonaire' );
-			
-			$options[0] = (array) $Bonaire_Options->default_options->{0};
-			$options[1] = (array) $Bonaire_Options->default_options->{1};
-			
-			add_option( 'bonaire_options', $options, '', true );
-			delete_transient( 'bonaire_reset_settings' );
+		// Gets the administrator role.
+		$role = get_role( 'administrator' );
+		
+		// If the acting user has admin rights, the capability gets added.
+		if ( ! empty( $role ) ) {
+			// Adds default settings data if there are no options stored yet.
+			if ( false === get_option( 'bonaire_options' ) || '1' === get_transient( 'bonaire_reset_settings' ) ) {
+				delete_option( 'bonaire_options' );
+				$Bonaire_Options = new AdminIncludes\Bonaire_Options( 'bonaire' );
+				
+				$options[0] = (array) $Bonaire_Options->default_options->{0};
+				$options[1] = (array) $Bonaire_Options->default_options->{1};
+				
+				add_option( 'bonaire_options', $options, '', true );
+				delete_transient( 'bonaire_reset_settings' );
+			}
 		}
 	}
